Using bootstrap to replace.
This site is hosted at the AWS S3 for a long time.
But using the Dropbox as the git repos.
Now switch to github again.
As I meet the data loose again, just because I reinstall the OS,
but forget to backup all the data needed.
And I reinstall the OS three times one day, different type of OS.
No way to find the data back.